pyproject.toml
[project]
name = "sheet-scraper"
version = "0.2.0"
description = "A tool to scrape product prices from supplier websites and update a Google Sheet."
authors = [
    { name = "Gemini", email = "gemini@google.com" },
]
dependencies = [
    "playwright",
    "undetected-playwright",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "pytest",
    "requests",
    "2captcha-python",
    "flask>=2.3.0",
]

[project.scripts]
sheet-scraper = "sheet_scraper.sheet_scraper:main"
